back to the previous navbar but changed a little
